refactor(admin): rename cohesion-calibration route to evaluator-calibration
Renames the frontend route /admin/cohesion-calibration → /admin/evaluator-calibration
and the backend Flask blueprint prefix /api/cohesion → /api/evaluator for parity
with the navbar Evaluator label landed in 51f88cb. Single user on the tool right
now, so bookmark / localStorage migration debt is acceptable.

- Directory rename via git mv: frontend/app/admin/cohesion-calibration → evaluator-calibration
- All client call sites in frontend/lib/api.ts swapped to /api/evaluator/*
- Internal import path in lib/formula-evaluator.ts updated to new dir
- NavBar dropdown + admin hub card href point at new route
- localStorage keys renamed (cohesion-calibration-* → evaluator-calibration-*);
  user loses prior test-history + lineup state on next visit (acceptable)
- Backend blueprint url_prefix updated; module docstring notes the prior path
- Integration test URLs updated; 7/7 pass

Intentionally untouched (per prior scope decision):
- Backend module file name (api/cohesion_calibration.py) and blueprint variable
  name (cohesion_calibration_bp) — internal identifiers
- services/cohesion_engine/ Layer — implementation engine name is the cohesion
  engine; Evaluator is the admin tool that calibrates it
- "Cohesion Score" product term in FAQ + Lab eval Surface — names the output
