+++ frontend/app/admin/evaluator-calibration/components/BellCurveCharts.tsx
