+++ frontend/app/admin/evaluator-calibration/components/FormulaPreview.tsx
