+++ frontend/app/admin/evaluator-calibration/components/PanelResizeHandle.tsx
