+++ frontend/app/admin/evaluator-calibration/components/PlayerInspection.tsx
