+++ frontend/app/admin/evaluator-calibration/components/ResultsPanel.tsx
