+++ frontend/app/admin/evaluator-calibration/components/WeightsEditor.tsx
